# Avoid colour space conversion when multiplying `SRGBColour` by white