Part of https://github.com/ppy/osu-framework/pull/6284.
Useful when computing `ColourInfo` of drawables which have white colour while applying colour to the parent (or changing parent's alpha).
Example would be applying colour/alpha to the tree of nested containers. In master some time will be spent computing colour on each level, but since by default they have white colour, we will basically passing it down to the children without conversions.
# Benchmark
master:
```
|                   Method |      Mean |    Error |   StdDev | Allocated |
|------------------------- |----------:|---------:|---------:|----------:|
|         MultiplyNonWhite | 140.09 ns | 0.385 ns | 0.341 ns |         - |
|            MultiplyWhite | 102.45 ns | 0.242 ns | 0.189 ns |         - |
| MultiplyWhiteWithOpacity | 102.51 ns | 0.221 ns | 0.173 ns |         - |
|         MultiplyConstOne |  96.63 ns | 0.176 ns | 0.165 ns |         - |
|      MultiplyConstNonOne |  97.18 ns | 0.269 ns | 0.225 ns |         - |
```
pr:
```
|                   Method |       Mean |     Error |    StdDev | Allocated |
|------------------------- |-----------:|----------:|----------:|----------:|
|         MultiplyNonWhite | 122.098 ns | 0.2430 ns | 0.1897 ns |         - |
|            MultiplyWhite |   6.129 ns | 0.0118 ns | 0.0110 ns |         - |
| MultiplyWhiteWithOpacity |   8.953 ns | 0.0491 ns | 0.0410 ns |         - |
|         MultiplyConstOne |   1.227 ns | 0.0057 ns | 0.0053 ns |         - |
|      MultiplyConstNonOne |  93.844 ns | 0.1871 ns | 0.1750 ns |         - |
```